Disallowed submitting improper card information

+++ routes.py
@@ -25,45 +25,49 @@ def newcard():
 @app.route("/newcard/send", methods=["POST"])
 def cardsend():
     cardname = request.form["cardname"]
-    finds = db.session.execute(text("SELECT DISTINCT L.name, L.id FROM users U, libraries L WHERE L.userid=U.id AND U.username=:username"), {"username":session["username"]}).fetchall()
-    print(finds , "kaikki on hienoa")
-    colours = ""
-    two_faced = False
-    power = None
-    toughness = None
-    inlibs = []
-    for value in request.form:
-        if value[:-1] == "colour":
-            colours += request.form[value]
-        elif value == "twofaced":
-            two_faced = request.form["twofaced"]
-        elif value == "power":
-            power = request.form["power"]
-        elif value == "toughness":
-            toughness = request.form["toughness"]
-        elif value != "cmc" and value != "rarity" and value != "cardname" and value != "libs":
-            inlibs.append((value, request.form[value]))
     cmc = request.form["cmc"]
     rarity = request.form["rarity"]
+    if len(cardname) > 0 and len(cmc) > 0 and len(rarity) > 0:
+        finds = db.session.execute(text("SELECT DISTINCT L.name, L.id FROM users U, libraries L WHERE L.userid=U.id AND U.username=:username"), {"username":session["username"]}).fetchall()
+        print(finds , "kaikki on hienoa")
+        colours = ""
+        two_faced = False
+        power = None
+        toughness = None
+        inlibs = []
+        for value in request.form:
+            if value[:-1] == "colour":
+                colours += request.form[value]
+            elif value == "twofaced":
+                two_faced = request.form["twofaced"]
+            elif value == "power":
+                power = request.form["power"]
+            elif value == "toughness":
+                toughness = request.form["toughness"]
+            elif value != "cmc" and value != "rarity" and value != "cardname" and value != "libs":
+                inlibs.append((value, request.form[value]))
+        
 
-    user = db.session.execute(text("SELECT id FROM users WHERE username=:username"), {"username":session["username"]}).fetchone()[0]
-    sql= """INSERT INTO cards (name, twofaced, colour, cmc, rarity, power, toughness, userid) 
-            VALUES (:cardname, :twofaced, :colours, :cmc, :rarity, :power, :toughness, :user)"""
-    db.session.execute(text(sql), {"cardname":cardname, "twofaced":two_faced, "colours":colours, "cmc":cmc, "rarity":rarity, "power":power, "toughness":toughness, "user":user})
-    db.session.commit()
-    userid = db.session.execute(text("""
-                            SELECT id FROM cards WHERE name=:cardname AND rarity=:rarity AND colour=:colours ORDER BY id DESC LIMIT 1"""
-                            ), {"cardname":cardname, "rarity":rarity, "colours":colours}).fetchone()[0]
-    if len(inlibs) != 0:
-        for i in inlibs:
-            print(i)
-            sql = "INSERT INTO cardlib (card, library, visible) VALUES (:userid, :library, True)"
-            db.session.execute(text(sql), {"userid":userid, "library":i[1]})
+        user = db.session.execute(text("SELECT id FROM users WHERE username=:username"), {"username":session["username"]}).fetchone()[0]
+        sql= """INSERT INTO cards (name, twofaced, colour, cmc, rarity, power, toughness, userid) 
+                VALUES (:cardname, :twofaced, :colours, :cmc, :rarity, :power, :toughness, :user)"""
+        db.session.execute(text(sql), {"cardname":cardname, "twofaced":two_faced, "colours":colours, "cmc":cmc, "rarity":rarity, "power":power, "toughness":toughness, "user":user})
+        db.session.commit()
+        userid = db.session.execute(text("""
+                                SELECT id FROM cards WHERE name=:cardname AND rarity=:rarity AND colour=:colours ORDER BY id DESC LIMIT 1"""
+                                ), {"cardname":cardname, "rarity":rarity, "colours":colours}).fetchone()[0]
+        if len(inlibs) != 0:
+            for i in inlibs:
+                print(i)
+                sql = "INSERT INTO cardlib (card, library, visible) VALUES (:userid, :library, True)"
+                db.session.execute(text(sql), {"userid":userid, "library":i[1]})
+        else:
+            sql = "INSERT INTO cardlib (card, library, visible) VALUES (:userid, 0, True)"
+            db.session.execute(text(sql), {"userid":userid})
+        db.session.commit()
+        return redirect("/")
     else:
-        sql = "INSERT INTO cardlib (card, library, visible) VALUES (:userid, 0, True)"
-        db.session.execute(text(sql), {"userid":userid})
-    db.session.commit()
-    return redirect("/")
+        return render_template("ohno.html", msg=2)
 
 @app.route("/signin", methods=["POST"])
 def signin():
@@ -121,39 +125,43 @@ def cardedit():
 def sendedit():
     card = request.form["card"]
     cardname = request.form["cardname"]
-    finds = db.session.execute(text("SELECT DISTINCT L.name, L.id FROM users U, libraries L WHERE L.userid=U.id AND U.username=:username"), {"username":session["username"]}).fetchall()
-    print(finds , "kaikki on hienoa")
-    colours = ""
-    two_faced = False
-    power = None
-    toughness = None
-    inlibs = []
-    for value in request.form:
-        if value[:-1] == "colour":
-            colours += request.form[value]
-        elif value == "twofaced":
-            two_faced = request.form["twofaced"]
-        elif value == "power":
-            power = request.form["power"]
-        elif value == "toughness":
-            toughness = request.form["toughness"]
-        elif value != "cmc" and value != "rarity" and value != "cardname" and value != "libs":
-            inlibs.append((value, request.form[value]))
     cmc = request.form["cmc"]
     rarity = request.form["rarity"] 
-    user = db.session.execute(text("SELECT id FROM users WHERE username=:username"), {"username":session["username"]}).fetchone()[0]
-    sql = """UPDATE cards SET name=:cardname, twofaced=:two_faced, colour=:colours,
-                 cmc=:cmc, rarity=:rarity, power=:power, toughness=:toughness WHERE id=:card"""
-    db.session.execute(text(sql), {"cardname":cardname, "two_faced":two_faced, "colours":colours, "cmc":cmc, "rarity":rarity, "power":power, "toughness":toughness, "card":card})
-    sql = """DELETE FROM cardlib WHERE card=:card"""
-    db.session.execute(text(sql), {"card":card})
-    if len(inlibs) != 0:
-        for i in inlibs:
-            print(i)
-            sql = "INSERT INTO cardlib (card, library, visible) VALUES (:userid, :library, True)"
-            db.session.execute(text(sql), {"userid":user, "library":i[1]})
+    if len(cardname) > 0 and len(cmc) > 0 and len(rarity) > 0:
+        finds = db.session.execute(text("SELECT DISTINCT L.name, L.id FROM users U, libraries L WHERE L.userid=U.id AND U.username=:username"), {"username":session["username"]}).fetchall()
+        print(finds , "kaikki on hienoa")
+        colours = ""
+        two_faced = False
+        power = None
+        toughness = None
+        inlibs = []
+        for value in request.form:
+            if value[:-1] == "colour":
+                colours += request.form[value]
+            elif value == "twofaced":
+                two_faced = request.form["twofaced"]
+            elif value == "power":
+                power = request.form["power"]
+            elif value == "toughness":
+                toughness = request.form["toughness"]
+            elif value != "cmc" and value != "rarity" and value != "cardname" and value != "libs":
+                inlibs.append((value, request.form[value]))
+        
+        user = db.session.execute(text("SELECT id FROM users WHERE username=:username"), {"username":session["username"]}).fetchone()[0]
+        sql = """UPDATE cards SET name=:cardname, twofaced=:two_faced, colour=:colours,
+                    cmc=:cmc, rarity=:rarity, power=:power, toughness=:toughness WHERE id=:card"""
+        db.session.execute(text(sql), {"cardname":cardname, "two_faced":two_faced, "colours":colours, "cmc":cmc, "rarity":rarity, "power":power, "toughness":toughness, "card":card})
+        sql = """DELETE FROM cardlib WHERE card=:card"""
+        db.session.execute(text(sql), {"card":card})
+        if len(inlibs) != 0:
+            for i in inlibs:
+                print(i)
+                sql = "INSERT INTO cardlib (card, library, visible) VALUES (:userid, :library, True)"
+                db.session.execute(text(sql), {"userid":user, "library":i[1]})
+        else:
+            sql = "INSERT INTO cardlib (card, library, visible) VALUES (:userid, 0, True)"
+            db.session.execute(text(sql), {"userid":user})
+        db.session.commit()
+        return redirect("/")
     else:
-        sql = "INSERT INTO cardlib (card, library, visible) VALUES (:userid, 0, True)"
-        db.session.execute(text(sql), {"userid":user})
-    db.session.commit()
-    return redirect("/")
+        return render_template("ohno.html", msg=2)
